# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GolfHero API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scores:
    get:
      operationId: getMyScores
      tags: [scores]
      summary: Get current user's scores (latest 5)
      responses:
        "200":
          description: List of scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"
    post:
      operationId: addScore
      tags: [scores]
      summary: Add a new score (rolling 5-score logic)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddScoreBody"
      responses:
        "201":
          description: Score added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Score for this date already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scores/{id}:
    put:
      operationId: updateScore
      tags: [scores]
      summary: Update an existing score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScoreBody"
      responses:
        "200":
          description: Updated score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteScore
      tags: [scores]
      summary: Delete a score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SUBSCRIPTIONS ─────────────────────────────────────────────────────────────
...
  /subscriptions/subscribe:
    post:
      operationId: subscribe
      tags: [subscriptions]
      summary: Create/activate a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscribeBody"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
...
  /draws:
    get:
      operationId: listDraws
      tags: [draws]
      summary: List all draws (published)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed]
      responses:
        "200":
          description: List of draws
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draw"
    post:
      operationId: createDraw
      tags: [draws]
      summary: Create a new draw (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDrawBody"
      responses:
        "201":
          description: Created draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
  /draws/current:
    get:
      operationId: getCurrentDraw
      tags: [draws]
      summary: Get the current active draw
      responses:
        "200":
          description: Current draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
        "404":
          description: No active draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /draws/{id}:
    get:
      operationId: getDraw
      tags: [draws]
      summary: Get a draw by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
  /draws/{id}/simulate:
    post:
      operationId: simulateDraw
      tags: [draws]
      summary: Simulate a draw (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateDrawBody"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawSimulationResult"
  /draws/{id}/publish:
    post:
      operationId: publishDraw
      tags: [draws]
      summary: Publish draw results (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Published draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"

  # ── PRIZES ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Score:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        score:
          type: integer
        scoreDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - score
        - scoreDate
        - createdAt
    AddScoreBody:
      type: object
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 45
        scoreDate:
          type: string
          format: date
      required:
        - score
        - scoreDate
    UpdateScoreBody:
      type: object
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 45
        scoreDate:
          type: string
          format: date
    SubscriptionStatus:
      type: object
      properties:
        id:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        plan:
          type: ["string", "null"]
          enum: [monthly, yearly, null]
        status:
          type: string
          enum: [active, inactive, cancelled, lapsed]
        renewalDate:
          type: ["string", "null"]
          format: date-time
        cancelledAt:
          type: ["string", "null"]
          format: date-time
        monthlyContributionAmount:
          type: number
      required:
        - status
        - monthlyContributionAmount
    SubscribeBody:
      type: object
      properties:
        plan:
          type: string
          enum: [monthly, yearly]
      required:
        - plan
...
    Draw:
      type: object
      properties:
        id:
          type: integer
        month:
          type: string
        year:
          type: integer
        status:
          type: string
          enum: [upcoming, active, simulated, published, completed]
        drawType:
          type: string
          enum: [random, algorithmic]
        winningNumbers:
          type: ["string", "null"]
        jackpotRolledOver:
          type: boolean
        jackpotAmount:
          type: number
        totalPool:
          type: number
        participantCount:
          type: integer
        publishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - month
        - year
        - status
        - drawType
        - jackpotRolledOver
        - jackpotAmount
        - totalPool
        - participantCount
        - createdAt
    CreateDrawBody:
      type: object
      properties:
        month:
          type: string
        year:
          type: integer
        drawType:
          type: string
          enum: [random, algorithmic]
      required:
        - month
        - year
        - drawType
    SimulateDrawBody:
      type: object
      properties:
        drawType:
          type: string
          enum: [random, algorithmic]
      required:
        - drawType
    DrawSimulationResult:
      type: object
      properties:
        drawId:
          type: integer
        winningNumbers:
          type: string
        matches5:
          type: integer
        matches4:
          type: integer
        matches3:
          type: integer
        prize5:
          type: number
        prize4:
          type: number
        prize3:
          type: number
      required:
        - drawId
        - winningNumbers
        - matches5
        - matches4
        - matches3
        - prize5
        - prize4
        - prize3
...